Change how we augmentate modules
