+++ src/types/modules/react-native-onyx.d.ts
@@ -1,0 +1,10 @@
+import {OnyxKey, OnyxCollectionKey, OnyxValues} from '../../ONYXKEYS';
+
+declare module 'react-native-onyx' {
+    // eslint-disable-next-line @typescript-eslint/consistent-type-definitions
+    interface CustomTypeOptions {
+        keys: OnyxKey;
+        collectionKeys: OnyxCollectionKey;
+        values: OnyxValues;
+    }
+}
